+++ Sources/PastePilotCLI/main.swift
@@ -1,0 +1,4 @@
+import Foundation
+import PastePilotCLIKit
+
+exit(PastePilotCLIApplication().run(arguments: CommandLine.arguments))